Added empty email parser function and referenced it into main
I added the file parse_one_email.py which contains the function parse_email. For the moment this function does not do anything, but it will contain the code to parse the content of a single email. 

I also added the needed code in main.py to use parse_email already in the intended place.