Added parsing email functionality to build a dataframe with the emails
I updated parse_one_email.py so that it now contains the functionalities to parse and return all the email metadata and the email body as a dictionary. With this new functionality I than updated main.py so that all the email metadata and bodies are stored in one unique pandas dataframe.