Updated email parsing capabilites and updated main
I updated parse_one_email.py so that now parse_email can return the body of multipart emails too and not only of simple emails. I also added a few extra passages to strip the email body of url addresses and html markdowns. In addition, I added some extra filtering procedures on the dataframe with the email so to make it more readable.