WIP - manipulating senders to create a unique labeling of email
I am manipulating a unique labeling in main.py to produce a unique labeling of the emails which allows to identify the hiring process from one company. I managed to isolate the sender email from other parts of the sender information so far, but I need to keep working on it.