Added clustering of emails using the From field of emails
I updated the main script so that now I cluster emails based on the contents of the From field in the emails. The clustering requires to preprocess the string in the From field with the text_utilities functions. Then I embed the From fields in a vector form, and finally I perform a distance-based hierarchical clustering. The code is still in a very coarse form, so I will need to clean it.