Polished clustering of emails using the from field
I have polished the clustering algorithm used to group together emails from the same company. To this end I had to tweak the height at which the hierarchical tree is cut, and I had to create a list of buzzwords that confounded the algorithm. The result is still improvable, but it works fine in most of the cases.