Added NMF decomposition to identify email bodies
I have updated the code in main.py and in the children funcions so that I can now identify topics with the Nonnegative matrix factorization. Well, it turns out that the non-negative mztrix factorization is too good since it identifies which companies I applied to, rather than a simple categorization of emails into "submission received", "submission rejected" and "feedback" that I was expecting. This reminds me that, possibly, a better idea is to use a Kmeans clustering from the TFIDF embedding. This way I can specify the number of clusters, and then try to see what each cluster represents.