Added Kmeans clustering
I added the kmeans clustering in main.py and the results are similar. I suspect that the words which characterizes the main three class of interest (received submission, rejected submission, feedback) are killed by some buzzwords. I guess that it's now time to build a supervised machine learning algorithm to identify these three classes from a selected set of emails, and then use that algorithm to classify the remaining emails.