Amended errors in stemming procedure
I updated the code because the addition of body preprocessing in the construction of the email dataframe destroyed the sentence structure of the email bodies. I solved the problem by including two fields in the email dataframe: "Bodies", which contains the original message, and "Stemmed Bodies", which contains the stemmed version stripped of all numbers and punctuation signs. I then updated main.py so that it looks for keywords from "Stemmed bodies", while it counts the sentence length from "Bodies"