WIP - Added code to compute the average rejection timespan
I added a few lines of code in main.py that allows me to compute the time it took to being rejected for each company where I submitted the candidacy. To this end, I consider as email from the same sender the emails that falls in the same cluster from the hierarchical tree classification I developed before. 

Note that not all the steps are implemented to get the final desired result.