Completed code to compute distribution of rejection days
I completed the code in main.py to compute the distribution of days it take to get a candidacy rejected. This code also required a minor update in build_email_dataframe.py to ensure that dates are expressed in datetime format. Finally, I also added a new figure in plots.py to save the final boxplot of the rejection time distribution.