Moved computation of rejection days to dedicated file
I moved all the code to compute the time it takes for an application to get rejected in the dedicated file compute_rejection_days.py. I also updated the code to perform such computation to save the automatic email grouping in an external file (not included for privacy reasons), so that I can check the goodness of the grouping.