Added dicussion about the rejection time in the latex report with extras
In this commit I mostly updated the latex document to include results about the rejection time of applications and how those relate to the company size. I also added some minor changes to the main.py code to get the information I needed to make the point in the latex report.